leapseconds: add parser for version 2 files
The latest file are going in version 2 with uint64 type used to store
all the timestamp information. Add parser for that version.
Also some changes to leap second structures - number of seconds is
signed value and should be read as signed. And also some optimisation
of skipping data without actual memory allocation. Tests are added