account holdover component for parts of the second
Summary: Previously holdover component was never accounted for WoU if the syncronisation was going smoothly. We have SYNC packets coming every second and the propability of getting at least 1 second between PHC ingestion and query very low. As we used integers to calculate the amount of seconds spent after the last ingestion, this value was always zero. But it does matter if we got a value 1 ms or 900ms after the last sync. Improve the quality using double as a value of seconds.

Reviewed By: t3lurid3

Differential Revision: D43472961

fbshipit-source-id: 8424791572adcf6d51284dd79f105ec9e0b93123