sptp use timer instead of ticker (#265)
Summary:
Pull Request resolved: https://github.com/facebook/time/pull/265

We want the servo adjustment interval to be as stable as possible, and 'ticker' in Go is too smart, it adjusts the interval if previous tick was too short or too long.

Also expose T1 to T4 in measurement result, just if we would like to use it/log it.

Reviewed By: leoleovich, vvfedorenko

Differential Revision: D43622416

fbshipit-source-id: 413c08f90e83b7f32176910a9472ae9fd539a7e4