Use poll for ERR queue only to read TX timestamp
Summary: The poll for POLLPRI can trigger on incoming packets too, not only on TX timestamp arrival on ERR queue. This races in sptp if the timestamp is delayed the worker could be woken up in poll() because of the answer from GM instead of TX timestamp. Fix it with waiting for ERR queue to wake worker up.

Reviewed By: leoleovich, abulimov

Differential Revision: D43846755

fbshipit-source-id: 36f1e7d70ab6fb1d6416e805c9d184064bf54193