feat: Add save config feature within the config subcommand (#315)
Summary:
Pull Request resolved: https://github.com/facebook/time/pull/315

Calnex builds a desired config by applying changes to the original config fetched from the device. If anything is updated, it will print the diff.

This feature allows users to show the entire config and not only the diff, even if there are no changes to the config. Which might be useful for debugging.

It is done by adding a --save <PATH> flag to Calnex's config subcommand. When --save <PATH> is provided, Calnex will dump the config file to <PATH>. Note that --save and --apply is mutually exclusive so either none is provided or one of the two.

Reviewed By: leoleovich

Differential Revision: D52538805

fbshipit-source-id: 25ba33331655e9fefdee9dfc57c8c453434c7470