fix: add trimming for TestSaveConfig config file assertion
Summary:
Save(...) calls .SaveTo(...) which for some reasons behave differently on different systems.

For example, when testing internally, it adds whitespace between the fields such that all "=" are aligned vertically. However when tests are executed in another environment (e.g. local with go test), it doesn't have this behaviour.

This diff removes the usage of .SaveTo(...) from the ini library and just do a normal bytes write to the filesystem for saving config.

Reviewed By: leoleovich

Differential Revision: D52625023