Improve the algo to calculate good mean frequency
Summary: Calculate good mean frequency based on the points where offset was within the range (new filter config option, default to 100ns). It takes frequency into account only when current and previous offsets were in the range. Only in this case we can say that the last frequency was not breaking PHC, or it was stable enough to be used in corner cases.

Reviewed By: abulimov

Differential Revision: D52788332

fbshipit-source-id: 5b2038c74a661b0332dacbfb7fb2d47c8301881c