push misbehaving GMs to the back of the list (#325)
Summary:
Pull Request resolved: https://github.com/facebook/time/pull/325

Credit to yarikk for suggesting this idea.
Because we trust local priority (when everything else is equal) we often end up jumping between time appliances.
For instance when 1st in the list is unavailable we jump to a 2nd. However, once the 1st is available we switch back to it. This creates unnecessary noise and in some cases leads to offset oscillations.

This diff will push 1st appliance to the end of the list and shift the rest of list up by 1.
As a result we will keep running on 2nd appliance until and never fallback to 1st.

Since we don't modify the config, restart will restore the order.

Reviewed By: t3lurid3, vvfedorenko

Differential Revision: D53191500

fbshipit-source-id: 5a4a62339484ea8a83f6969f273e8654354cf371