format antenna delay to microseconds if necessary
Summary:
Temporarily add a hack to match Calnex microseconds format.

Apparently, the device is trying to be smart and after setting the values returns `antenna_delay=3.042 us`, which in turn triggers our automation to set the value again and again.
It's a bit painful to match this string especially because golang is using proper unicode characters https://go.dev/play/p/_awe7ecCSuq
```
3.042µs
```

Reviewed By: pmazzini

Differential Revision: D54415146

fbshipit-source-id: 08860676ec3c42e84578dfde2b998ef3c1456ff6