calculate W immediately
Summary:
Currently we are waiting 200 seconds for fbclock to restart. All this time we get high WOU or even an error if host was rebooted.
This is due to the fact we first calculate `M` and then `W`.
However, we can keep calculating `W` with the data we have.
We will still complain if `M` has less than min data points, but by the time we have it at 100 we will also have 100 `W` and ready to serve the client.
This will save us 50% of start time.

Reviewed By: abulimov

Differential Revision: D55199478

fbshipit-source-id: dcf64ae95312f104d3257c2872e04b7047e026ed