Add static lib target (#374)
Summary:
Pull Request resolved: https://github.com/facebook/time/pull/374

Previous changes to conda build caused build breakages due to fbclock.so not being copied correctly. One of the suggestions was to make it static lib that further links to libnccl. So that all the upper layers only have to worry about managing libnccl.so, as they already do and not concern with managing other dependent libs.

In this take2, introduce static lib target for fbclock, that will be further be linked to libnccl.so in subsequent changes.

Reviewed By: leoleovich

Differential Revision: D58930591

fbshipit-source-id: c6bbaa6e1e6c050b302b11b3a7d682e799d049b9