Return the correct reply packet type of Chrony reply packets
Changes the return value function `GetType` to be of type `ReplyType`
instead of type `PacketType`. The latter merely captures whether a given
packet is a request or a reply, which is thus redundant given that
`GetType()` operates on a `ResponsePacket` that by definition is always
a reply (same argument for `ReplyHead`).

The comment of `GetType` suggests that it returns the "reply packet
type", which is precisely what `type ReplyType` captures according to
its documentation. We therefore assume that this to be a bug.