Return the correct reply packet type of Chrony reply packets (#376)
Summary:
Changes the return value function `GetType` to be of type `ReplyType` instead of type `PacketType`. The latter merely captures whether a given packet is a request or a reply, which is thus redundant given that `GetType()` operates on a `ResponsePacket` that by definition is always a reply (same argument for `ReplyHead`).

The comment of `GetType` suggests that it returns the "reply packet type", which is precisely what `type ReplyType` captures according to its documentation. We therefore assume that this to be a bug.

Also now exports the different `ReplyType` constants such that they can be used along with what `GetType()` now returns.

Pull Request resolved: https://github.com/facebook/time/pull/376

Reviewed By: leoleovich

Differential Revision: D60377340

Pulled By: abulimov

fbshipit-source-id: 0661ce95e5401b2476fb2dc8bf736864d72b5d24