phc: introduce phc.Device abstraction
Summary:
Manage the number of package-level functions of mixed purpose, group the ones which deal with PHC devices as methods of a the new Device type.

Reasoning: with planned addition of the functionality of *testptp* to the `phc` package and utils, the more of the similar methods are upcoming, and if not addressed early, the set of package-level functions is going to oganically grow even more crowded.

Reviewed By: abulimov

Differential Revision: D60456794