add support for reporting cumulative switch latency in path
Summary:
Per hop CF values can be negative even when PTP TCs are operating correctly due to the nature in how we calculate these values. A brief period of congestion on the prior hop can cause the per CF value to be negative (tho with a minimum number of packets per hop > 1 the odds of this decreases).

Running `ziffy` with `-raw` returns the actual CF value encoded by each TC in the path.

Reviewed By: abulimov

Differential Revision: D61275719

fbshipit-source-id: c6a9eea1d84d7d9b3979aeb9b3a3892a4ec0c3af