custom UDPConn implementation (#387)
Summary:
Pull Request resolved: https://github.com/facebook/time/pull/387

We already mostly just used FD from the connection to set up the timestamping and such.
Remove the middle man and create socket and bind it ourselves.

This allows us to have full control over the connection, such as using SO_REUSEPORT (see diff up stack), which is impossible with standard library.

As a bonus, with this change sptp/ptping will work with IPv4 environment thanks to new config option and flag that allows to specify IP to bind to.

Reviewed By: leoleovich

Differential Revision: D61796514