ptpcheck: phc: add clock setting function
Summary:
Enable the `ptpcheck phc` command to set the PTP clock to a given value.

The new flag, `-s`, takes a Unix timestamp as an argument, and writes it to the PTP device. The resolution is thus limited to whole seconds, so the clock then requires a more precise tuning – achievable, for example, with the sptp program.

Differential Revision: D64134612