ptpcheck: phc: add clock setting function (#402)
Summary:
Pull Request resolved: https://github.com/facebook/time/pull/402

Enable the `ptpcheck phc` command to set the PTP clock to a given value.

The new flag, `-s`, takes a Unix timestamp as an argument, and writes it to the PTP device. The resolution is thus limited to whole seconds, so the clock then requires a more precise tuning – achievable, for example, with the sptp program.

Reviewed By: abulimov

Differential Revision: D64134612

fbshipit-source-id: 5f89bb96a5539760476a72d11cc24191b6184ab7