reorder timestamp enablement
Summary:
We see the situation where on Broadcom hosts with a lot of traffic we spend up to **20% of entire CPU** in softirq due to spin lock implementation (S456418).
In short - if supported - we need to try to enable PTPv2L4 timestamping first and only if it's not supported timestamp all.
In the same diff removing PTPv2L2 timestamping as we don't do PTP over ethernet.

Reviewed By: vvfedorenko

Differential Revision: D64391681