reorder timestamp enablement (#405)
Summary:
Pull Request resolved: https://github.com/facebook/time/pull/405

We see the situation where on Broadcom hosts with a lot of traffic we spend up to **20% of entire CPU** in softirq due to spin lock implementation (S456418).
In short - if supported - we need to try to enable PTPv2L4 timestamping first and only if it's not supported timestamp all.
In the same diff removing PTPv2L2 timestamping as we don't do PTP over ethernet.

Reviewed By: abulimov, vvfedorenko

Differential Revision: D64391681

fbshipit-source-id: 5f9c6294874dd65f2966fcaca55dd6164b4c32ce