reset FirstUpdate flag on servo after we enter locked state (#412)
Summary:
Pull Request resolved: https://github.com/facebook/time/pull/412

Currently we can step at any point in the future if something unexpected happens.
We never reset the flag, so techically if the servo was reset we could step multiple times!

With this change, once servo is in S2 (locked), we unset the flag. This ensures that we can only step once at the beginning of SPTP lifetime.

As a bonus I've added a log message to make it more clear when we step the clock.

Reviewed By: leoleovich

Differential Revision: D64602914

fbshipit-source-id: cd70487124080d2093006b7dd2f176980e0cd843