Initial implementation of asymmetry correction (#449)
Summary:
Pull Request resolved: https://github.com/facebook/time/pull/449

WHAT?

Implements asymmetry correction by changing the port of GMs which have offset over a configurable threshold. If chaning GMs port does not solve, eventually currently selected GM is considered to be asymmetric and it's port is changed.

WHY?
Lower offset on the fleet, alowing us to tighten our SLA.
This first version allows us to test values for the configs on a subset of the fleet.

Considerations:
Why this approach?
 - Alternative approach of changing only selected GM and only when all other GMs are bad will be tested comparatively.
 - Current approach might lead to more port changes, but also seems promising in terms of leading to overall lower offsets

Reviewed By: t3lurid3, vvfedorenko

Differential Revision: D71034365

fbshipit-source-id: cbc7eea91fabd8bee75f2854d03e9b6fde32c3d0