Add option to use SPTP over PTP for serf-based linearizability test
Summary:
WHAT?
Allows us to use sptp instead of PTP for linearizability tests

WHY?
Subscription means additional load on time servers, and time servers may not respond to the subscription timely, which would generate a need for further complexity in an already complex test scenario. The usage of SPTP reduces load, simplifies the testing flow and makes it more reliable.

Reviewed By: vvfedorenko

Differential Revision: D73510418

fbshipit-source-id: bbcc5cb613df30cf4b341a189d8bd8f26bf831a7