improve FW versioning logic
Summary: The latest update to calnex firmware tool was removing the first part of the version as it's HW specific. That's correct for OSS variant were we provide FW as a file. But the logic was not corrected for internal usage where we fetch version from netcode. Do the needed modification

Reviewed By: t3lurid3

Differential Revision: D73854347

fbshipit-source-id: d58614dcc185f5e5f6a5016d4b72035e0526d4e2