Calnex: Don't attempt to set settings that aren't present to start with (#459)
Summary:
edit configs for base config tests to mirror unlicensed channels

Resolves issues with newer units that don't have port 2 licenses enabled as there are no physical ports on the device.
Adjust code to only set a setting if it is present in the settings returned from the Calnex device when settings are already present from a getsettings call.
If a setting in not present in a getsettings api call then it will not be possible to apply it to a unit.

Moved some of the unit tests to use channel C instead of channel A. Channel A is not an available clock channel on most instruments as the packet channel uses A and B
edit configs for base config tests to mirror unlicensed channels **[this is an additional change over the last PR that came about after merging changes]**

Pull Request resolved: https://github.com/facebook/time/pull/459

Test Plan:
apply config to box with no channel 1 Licenses.
(doing the same for channel2 doesn't exercise the new code as the current fix put in for this successfully prevents attempting to set channel 2 settings)
A measurement starts as expected in this case
![image](https://github.com/user-attachments/assets/32e7a014-8b3d-44af-bd82-be684151c9c3)
Output if there is no virtual channel license
![image](https://github.com/user-attachments/assets/78f0499f-bb12-471d-94be-f8d841f470cc)

Reviewed By: t3lurid3

Differential Revision: D74585364

Pulled By: vvfedorenko

fbshipit-source-id: e24ce67eb217f73ddc905af22bfe3f709e51101e