make a step in case of stalled CPU
Summary: When sptp is not scheduled for more than a minute it means that the host was overhelmed. Mellanox driver used system work queue to keep time running on old kernels which also can be stalled. That's one of the reasons of huge offset on CX4 and CX6. There is no real reason to do normal sync and we can jump clocks to converge faster.

Reviewed By: t3lurid3

Differential Revision: D74723431

fbshipit-source-id: 9f9a880acc454002776c34bbc96909440fe3385e