ptp4u: ensure `workerOffset` is always initialized to zero for each new packet (#472)
Summary:
Fixes https://github.com/facebook/time/issues/471

A bit defence in depth to ensure `workerOffset` is always initialized to zero for each new packet processed. The sptp client sets the value to zero by default, but It'd be a good practice to just reset the value if the client doesn't provide it for whatever reason.

Pull Request resolved: https://github.com/facebook/time/pull/472

Test Plan: I didn't write any unit tests to check for this.

Reviewed By: pmazzini

Differential Revision: D79648907

Pulled By: vvfedorenko

fbshipit-source-id: 6433f92ba8f6c18d1b226bfc3b24109a479208c9