open PHC devices with O_RDWR
Summary:
both devices are configured during the run of ts2phc, so they both have to be opened with write permissions. Also, the time for the pps in signal should be the represented as Unix() not Second() (which returns the second in the minute of the timestamp).
Lastly, if PTP_PEROUT_DUTY_CYCLE didn't work, the durty cycle srtucture has to be cleaned out to have kernel accept the change

Reviewed By: deathowl

Differential Revision: D80630210

fbshipit-source-id: c087d5ed504ba5a9350969fb31ea51714c96d5f6