Calnex: do not try to change settings for channels that aren't installed (#477)
Summary:
In our upcoming firmware release we have had to change how our settings files are applied and ignore any settings that are for channels that are not installed

This results in the GO tool constantly seeing some settings that it thinks are wrong for channels that are not installed so constantly stopping a measurement and applying a new settings file where some of the settings are being ignored.

These changes check for the installed setting for each channel being 1 before allowing changes to the expected settings for comparison before upload

There is a simpler way to mitigate this which is to remove the references to channels A and B so those channels are never set as they are the main culprits, but it is less robust, and would be less useful for anyone else using the applicaiton.

Pull Request resolved: https://github.com/facebook/time/pull/477

Test Plan: I have added to the unit tests. I can do some more testing if you are happy with this as the solution and not just removing channel A/B references.

Reviewed By: vvfedorenko

Differential Revision: D84908650

Pulled By: leoleovich

fbshipit-source-id: dce8ed08e213535f8fb837fbc5b59e8278b86d07