Open source fbclock command in ptpcheck
Summary:
We previously had an internal fbclock command implementation in fbcode//ptp/ptpcheck that was not part of the open source time library. This change migrates the fbclock functionality to the open source ptpcheck tool in fbcode//time/cmd/ptpcheck, making it available for external use.

This change adds the fbclock subcommand to the open source ptpcheck tool, along with the necessary build dependencies.

Differential Revision: D88368651