tweak linearizability testing logic
Summary:
Use case where an `sptp` client does not have other GMs to compare with active GM was reporting as linearizability test failure which is not intended behaviour. Not being able to compare time delta between the client and other GMs should not mean there is an issue with linearizability. It means we don't know at this time.

Linearizability Test Logic is:
1. Test each other GMs that the client is configured to use (but are not the active GM for the client).
2. The other GMs must be directly locked to a highly accurate, primary reference time source (or we don't test them)
3. If a test fails because say the time delta > the  configured `LinearizabilityTestMaxGMOffset` then record the failure
4. If all tests fail i.e. all other GMs that are tested (and there is at least one GM to test) then that will bump the `ptp.fbclock.clock.linearizability.failed` metric

Reviewed By: virtyaluk

Differential Revision: D88433729

fbshipit-source-id: a671044c65d38a3f19ab8860b06a7750828a2336