Open source fbclock command in ptpcheck (#481)
Summary:
Pull Request resolved: https://github.com/facebook/time/pull/481

We previously had an internal fbclock command implementation in fbcode//ptp/ptpcheck that was not part of the open source time library. This change migrates the fbclock functionality to the open source ptpcheck tool in fbcode//time/cmd/ptpcheck, making it available for external use.

This change adds the fbclock subcommand to the open source ptpcheck tool, along with the necessary build dependencies.

Reviewed By: leoleovich

Differential Revision: D88368651